backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==23.0.0
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.1.1

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
google-auth==2.35.0
google-auth-oauthlib==1.2.1

# HTTP / async
httpx==0.27.2
aiofiles==24.1.0

# Storage
boto3==1.35.23          # AWS S3
google-cloud-storage==2.18.2  # GCS

# Background tasks
celery[redis]==5.4.0
apscheduler==3.10.4
redis==5.1.0

# ML / OCR
pytesseract==0.3.13
Pillow==10.4.0
numpy==1.26.4
# tflite-runtime==2.14.0   # uncomment for edge ML

# PDF generation
reportlab==4.2.4

# Observability
structlog==24.4.0
prometheus-client==0.21.0
sentry-sdk[fastapi]==2.14.0

# Security
cryptography==43.0.1

# Utilities
python-dateutil==2.9.0
babel==2.16.0
pycountry==23.12.11